third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-panel-structure.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/elements-test.js"></script>
<script>

function test() {
  // Warm up highlighter module.
  runtime.loadModulePromise('source_frame').then(function() {
    ElementsTestRunner.expandElementsTree(step1);
  });

  function step1() {
    ElementsTestRunner.dumpElementsTree();
    TestRunner.completeTest();
  }
}

function prepareTest()
{
    document.querySelector("#control-character").textContent = "\ufeff\u0093";
    runTest();
}

</script>
</head>

<body onload="prepareTest()">
<p>
Tests that elements panel shows DOM tree structure.
</p>

<div id="level1">
    <div id="level2">"Quoted Text". Special&nbsp;characters: &gt;&lt;&quot;&apos;&nbsp;&ensp;&emsp;&thinsp;&#8202;&#8203;&zwnj;&zwj;&lrm;&rlm; &#8234;&#8235;&#8236;&#8237;&#8238;&shy;<div id="level3"></div>
    </div>
</div>
<div id="control-character"></div>

</body>
</html>
