third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-panel-structure-expected.txt

Tests that elements panel shows DOM tree structure.

"Quoted Text". Special characters: ><"'     ​‌‍‎‏ ‪‫‬‭‮­
﻿
  <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
- <html>
    - <head>
          <script src="../../inspector/inspector-test.js"></script>
          <script src="../../inspector/elements-test.js"></script>
        - <script>
              \nfunction test() {\n  // Warm up highlighter module.\n  runtime.loadModulePromise('source_frame').then(function() {\n    ElementsTestRunner.expandElementsTree(step1);\n  });\n\n  function step1() {\n    ElementsTestRunner.dumpElementsTree();\n    TestRunner.completeTest();\n  }\n}\n\nfunction prepareTest()\n{\n    document.querySelector("#control-character").textContent = "\ufeff\u0093";\n    runTest();\n}\n\n
          </script>
      </head>
    - <body onload="prepareTest()">
          <p>\nTests that elements panel shows DOM tree structure.\n</p>
        - <div id="level1">
            - <div id="level2">
                  ""Quoted Text". Special&nbsp;characters: ><"'&nbsp;&ensp;&emsp;&thinsp;&#8202;&#8203;&zwnj;&zwj;&lrm;&rlm; &#8234;&#8235;&#8236;&#8237;&#8238;&shy;"
                  <div id="level3"></div>
              </div>
          </div>
          <div id="control-character">&#65279;&#147;</div>
      </body>
  </html>
